Issue #147: Use 64-bit Java type for numbers in the debug protocol

+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/Breakpoint.java
@@ -22,7 +22,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  private Integer id;
+  private Long id;
   
   /**
    * If true breakpoint could be set (but not necessarily at the desired location).
@@ -50,21 +50,21 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  private Integer line;
+  private Long line;
   
   /**
    * An optional start column of the actual range covered by the breakpoint.
    * <p>
    * This is an optional property.
    */
-  private Integer column;
+  private Long column;
   
   /**
    * An optional end line of the actual range covered by the breakpoint.
    * <p>
    * This is an optional property.
    */
-  private Integer endLine;
+  private Long endLine;
   
   /**
    * An optional end column of the actual range covered by the breakpoint. If no end line is given, then the end
@@ -72,7 +72,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  private Integer endColumn;
+  private Long endColumn;
   
   /**
    * An optional unique identifier for the breakpoint.
@@ -80,7 +80,7 @@ public class Breakpoint {
    * This is an optional property.
    */
   @Pure
-  public Integer getId() {
+  public Long getId() {
     return this.id;
   }
   
@@ -89,7 +89,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  public void setId(final Integer id) {
+  public void setId(final Long id) {
     this.id = id;
   }
   
@@ -155,7 +155,7 @@ public class Breakpoint {
    * This is an optional property.
    */
   @Pure
-  public Integer getLine() {
+  public Long getLine() {
     return this.line;
   }
   
@@ -164,7 +164,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  public void setLine(final Integer line) {
+  public void setLine(final Long line) {
     this.line = line;
   }
   
@@ -174,7 +174,7 @@ public class Breakpoint {
    * This is an optional property.
    */
   @Pure
-  public Integer getColumn() {
+  public Long getColumn() {
     return this.column;
   }
   
@@ -183,7 +183,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  public void setColumn(final Integer column) {
+  public void setColumn(final Long column) {
     this.column = column;
   }
   
@@ -193,7 +193,7 @@ public class Breakpoint {
    * This is an optional property.
    */
   @Pure
-  public Integer getEndLine() {
+  public Long getEndLine() {
     return this.endLine;
   }
   
@@ -202,7 +202,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  public void setEndLine(final Integer endLine) {
+  public void setEndLine(final Long endLine) {
     this.endLine = endLine;
   }
   
@@ -213,7 +213,7 @@ public class Breakpoint {
    * This is an optional property.
    */
   @Pure
-  public Integer getEndColumn() {
+  public Long getEndColumn() {
     return this.endColumn;
   }
   
@@ -223,7 +223,7 @@ public class Breakpoint {
    * <p>
    * This is an optional property.
    */
-  public void setEndColumn(final Integer endColumn) {
+  public void setEndColumn(final Long endColumn) {
     this.endColumn = endColumn;
   }
