+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/EvaluateResponse.java
@@ -42,7 +42,7 @@ public class EvaluateResponse {
    * variablesReference to the VariablesRequest.
    */
   @NonNull
-  private Integer variablesReference;
+  private Long variablesReference;
   
   /**
    * The number of named child variables.
@@ -51,7 +51,7 @@ public class EvaluateResponse {
    * <p>
    * This is an optional property.
    */
-  private Integer namedVariables;
+  private Long namedVariables;
   
   /**
    * The number of indexed child variables.
@@ -60,7 +60,7 @@ public class EvaluateResponse {
    * <p>
    * This is an optional property.
    */
-  private Integer indexedVariables;
+  private Long indexedVariables;
   
   /**
    * The result of the evaluate request.
@@ -122,7 +122,7 @@ public class EvaluateResponse {
    */
   @Pure
   @NonNull
-  public Integer getVariablesReference() {
+  public Long getVariablesReference() {
     return this.variablesReference;
   }
   
@@ -130,7 +130,7 @@ public class EvaluateResponse {
    * If variablesReference is > 0, the evaluate result is structured and its children can be retrieved by passing
    * variablesReference to the VariablesRequest.
    */
-  public void setVariablesReference(@NonNull final Integer variablesReference) {
+  public void setVariablesReference(@NonNull final Long variablesReference) {
     this.variablesReference = variablesReference;
   }
   
@@ -142,7 +142,7 @@ public class EvaluateResponse {
    * This is an optional property.
    */
   @Pure
-  public Integer getNamedVariables() {
+  public Long getNamedVariables() {
     return this.namedVariables;
   }
   
@@ -153,7 +153,7 @@ public class EvaluateResponse {
    * <p>
    * This is an optional property.
    */
-  public void setNamedVariables(final Integer namedVariables) {
+  public void setNamedVariables(final Long namedVariables) {
     this.namedVariables = namedVariables;
   }
   
@@ -165,7 +165,7 @@ public class EvaluateResponse {
    * This is an optional property.
    */
   @Pure
-  public Integer getIndexedVariables() {
+  public Long getIndexedVariables() {
     return this.indexedVariables;
   }
   
@@ -176,7 +176,7 @@ public class EvaluateResponse {
    * <p>
    * This is an optional property.
    */
-  public void setIndexedVariables(final Integer indexedVariables) {
+  public void setIndexedVariables(final Long indexedVariables) {
     this.indexedVariables = indexedVariables;
   }
