+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/Scope.java
@@ -28,7 +28,7 @@ public class Scope {
    * VariablesRequest.
    */
   @NonNull
-  private Integer variablesReference;
+  private Long variablesReference;
   
   /**
    * The number of named variables in this scope.
@@ -37,7 +37,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  private Integer namedVariables;
+  private Long namedVariables;
   
   /**
    * The number of indexed variables in this scope.
@@ -46,7 +46,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  private Integer indexedVariables;
+  private Long indexedVariables;
   
   /**
    * If true, the number of variables in this scope is large or expensive to retrieve.
@@ -66,28 +66,28 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  private Integer line;
+  private Long line;
   
   /**
    * Optional start column of the range covered by this scope.
    * <p>
    * This is an optional property.
    */
-  private Integer column;
+  private Long column;
   
   /**
    * Optional end line of the range covered by this scope.
    * <p>
    * This is an optional property.
    */
-  private Integer endLine;
+  private Long endLine;
   
   /**
    * Optional end column of the range covered by this scope.
    * <p>
    * This is an optional property.
    */
-  private Integer endColumn;
+  private Long endColumn;
   
   /**
    * Name of the scope such as 'Arguments', 'Locals'.
@@ -111,7 +111,7 @@ public class Scope {
    */
   @Pure
   @NonNull
-  public Integer getVariablesReference() {
+  public Long getVariablesReference() {
     return this.variablesReference;
   }
   
@@ -119,7 +119,7 @@ public class Scope {
    * The variables of this scope can be retrieved by passing the value of variablesReference to the
    * VariablesRequest.
    */
-  public void setVariablesReference(@NonNull final Integer variablesReference) {
+  public void setVariablesReference(@NonNull final Long variablesReference) {
     this.variablesReference = variablesReference;
   }
   
@@ -131,7 +131,7 @@ public class Scope {
    * This is an optional property.
    */
   @Pure
-  public Integer getNamedVariables() {
+  public Long getNamedVariables() {
     return this.namedVariables;
   }
   
@@ -142,7 +142,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  public void setNamedVariables(final Integer namedVariables) {
+  public void setNamedVariables(final Long namedVariables) {
     this.namedVariables = namedVariables;
   }
   
@@ -154,7 +154,7 @@ public class Scope {
    * This is an optional property.
    */
   @Pure
-  public Integer getIndexedVariables() {
+  public Long getIndexedVariables() {
     return this.indexedVariables;
   }
   
@@ -165,7 +165,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  public void setIndexedVariables(final Integer indexedVariables) {
+  public void setIndexedVariables(final Long indexedVariables) {
     this.indexedVariables = indexedVariables;
   }
   
@@ -210,7 +210,7 @@ public class Scope {
    * This is an optional property.
    */
   @Pure
-  public Integer getLine() {
+  public Long getLine() {
     return this.line;
   }
   
@@ -219,7 +219,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  public void setLine(final Integer line) {
+  public void setLine(final Long line) {
     this.line = line;
   }
   
@@ -229,7 +229,7 @@ public class Scope {
    * This is an optional property.
    */
   @Pure
-  public Integer getColumn() {
+  public Long getColumn() {
     return this.column;
   }
   
@@ -238,7 +238,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  public void setColumn(final Integer column) {
+  public void setColumn(final Long column) {
     this.column = column;
   }
   
@@ -248,7 +248,7 @@ public class Scope {
    * This is an optional property.
    */
   @Pure
-  public Integer getEndLine() {
+  public Long getEndLine() {
     return this.endLine;
   }
   
@@ -257,7 +257,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  public void setEndLine(final Integer endLine) {
+  public void setEndLine(final Long endLine) {
     this.endLine = endLine;
   }
   
@@ -267,7 +267,7 @@ public class Scope {
    * This is an optional property.
    */
   @Pure
-  public Integer getEndColumn() {
+  public Long getEndColumn() {
     return this.endColumn;
   }
   
@@ -276,7 +276,7 @@ public class Scope {
    * <p>
    * This is an optional property.
    */
-  public void setEndColumn(final Integer endColumn) {
+  public void setEndColumn(final Long endColumn) {
     this.endColumn = endColumn;
   }
